tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.double.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.create2.double</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.create2.double</h1>
<p class="desc">createImageData(w, h) double is converted to long</p>


<script>
var t = async_test("createImageData(w, h) double is converted to long");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var imgdata1 = ctx.createImageData(10.01, 10.99);
  var imgdata2 = ctx.createImageData(-10.01, -10.99);
  _assertSame(imgdata1.width, 10, "imgdata1.width", "10");
  _assertSame(imgdata1.height, 10, "imgdata1.height", "10");
  _assertSame(imgdata2.width, 10, "imgdata2.width", "10");
  _assertSame(imgdata2.height, 10, "imgdata2.height", "10");
  t.done();

});
</script>
